contact/serializers.py
# contact/serializers.py
from rest_framework import serializers
from .models import Contact, PhoneNumber


class ContactSerializer(serializers.ModelSerializer):
    class Meta:
        model = Contact
        fields = '__all__'
    

class PhoneNumberSerializer(serializers.ModelSerializer):
    class Meta:
        model = PhoneNumber
        fields = '__all__'
