requirements.txt
asgiref==3.5.2
autopep8==1.4.4
backports.zoneinfo==0.2.1
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==5.0.0
Markdown==3.4.1
pycodestyle==2.5.0
pytz==2019.3
sqlparse==0.3.0
zipp==3.10.0
